Tests: Move blocks raw handling tests to `test/integration` folder (#6731)
* Tests: Move integration tests to core blocks module

* Tests: Move block raw handling tests to integration folder
