+++ test/integration/fixtures/classic-in.html
