+++ test/integration/fixtures/one-image-in.html
