+++ test/integration/fixtures/one-image-out.html
